window-x11: Focus the default window with delay while waiting for take-focus
When requesting to a take-focus window to acquire the input, the client may or
may not respond with a SetInputFocus (this doesn't happen for no-input gtk
windows in fact [to be fixed there too]), in such case we were unsetting the
focus while waiting the reply.

In case the client won't respond, we wait for a small delay (set to 250 ms) for
the take-focus window to grab the input focus before setting it to the default
window.

Added a test for this behavior and for the case in which a window takes the
focus meanwhile we're waiting to focus the default window.

(cherry picked from commit f71151a5dd990d935f3fbb39451f9b41f640b625)

Bug: https://gitlab.gnome.org/GNOME/mutter/merge_requests/307
Origin: upstream, 3.30.3, commit:8abd1b00926534fb9a34638c89cbb41b37f0e078

Gbp-Pq: Name window-x11-Focus-the-default-window-with-delay-while-wait.patch