window-actor: Don't show actor until meta_window_actor_show()
By default clutter will show an actor as it is added to a parent. This
means that after we create the window actor, when it's added to the
window group, we implicitly show it. What we really want is to not show
it until the window is supposed to be shown, which happens when
meta_window_actor_show() is called, as showing prior to that, could
cause issues.

Avoid the implicit show by setting the "show-on-set-parent" property on
the window actor to `FALSE` on window actor construction.

Forwarded: https://gitlab.gnome.org/GNOME/mutter/merge_requests/1066
Origin: upstream, 3.34.5, commit:011ff15603d6f1a25e7b4bb79bfc2bcdcef486e5

Gbp-Pq: Name window-actor-Don-t-show-actor-until-meta_window_actor_sho.patch