clutter/actor: Fix pick when actor is not allocated
When selecting the pick regions for an actor we were not considering
whether the actor was allocated and that was causing issues where the
preferred width/height of the actor was used when deciding whether
the actor should be considered as a pick target.

Check if the actor has a valid allocation, in addition to being mapped
and being in pick mode, in clutter_actor_should_pick_paint().

Forwarded: https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1169
Applied-upstream: 3.36.2, commit:82f3bdd14e0081dff60e9fed51376fc4cbf8b201

Gbp-Pq: Name clutter-actor-Fix-pick-when-actor-is-not-allocated.patch