wayland/xdnd: Add error traps around Xdnd* IPC
Make all of them spew criticals, except for XdndLeave as it's feasible
to expect the window we are sending the event to did disappear in the
way (eg. if the window is destroyed while the DnD operation is ongoing
and the pointer is over the window).

Bug: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/2590
Forwarded: https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1184
Origin: upstream, 3.36.2

Gbp-Pq: Name wayland-xdnd-Add-error-traps-around-Xdnd-IPC.patch