window: Do not go past size hints on resize
On interactive resize, mutter calculates the difference in size based on
the pointer location and relies on window constraints to ensure the
minimum size is honored.

Wayland however does asynchronous window configuration, meaning that not
checking for size hints early enough may lead to the window moving as
the locations was initially computed on a size which will be invalidate
by the client eventually.

Make sure to respect the client size hint on update_resize() so that we
don't end up with a window moving unexpectedly when the client
eventually acked the configuration.

https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1495

(cherry picked from commit 03c69ed8cf8059d98089929b9d79b92df0167fbe)

Bug-Ubuntu: https://bugs.launchpad.net/ubuntu/+source/mutter/+bug/1878293

Gbp-Pq: Name window-Do-not-go-past-size-hints-on-resize.patch