mutter-all.test: Run tests in a environment with display and bus set
The mutter tests require to run in a valid environment where a display
is available and a session bus, however currently we rely on the current
environment, and this may lead to unexpected behaviors.

So let's just ensure that a display is running through xvfb-run and
that a session bus is running in a temporary directory.

We also ensure to use the gsettings memory backend, even because by
setting TestEnvironment we ensure that no other env variable is leaked
to the test.

Origin: https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1876

Gbp-Pq: Name mutter-all.test-Run-tests-in-a-environment-with-display-a.patch