tests: Make it easier to keep track of where we are
The monitor orientation tests do a lot of things in sequence. Replace
some of the comments with g_test_message() so that the log from a failed
test gives us a better idea of how far we got.

Signed-off-by: Simon McVittie <smcv@debian.org>
Forwarded: https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/2049
Applied-upstream: 42, commit:838d5656c347d040892347babea26c71400caed6

Gbp-Pq: Name tests-Make-it-easier-to-keep-track-of-where-we-are.patch