workspace: Don't crash on invalid argument to meta_workspace_index
This is no longer strictly necessary after #2559 was fixed in 44.beta,
but could add some extra robustness by turning a crash into a somewhat
graceful check failure.

TODO: If the upstream merge request is accepted, we can keep applying
this until we upgrade to a version that includes it; but if the upstream
merge request is rejected, we should drop this.

Bug: https://gitlab.gnome.org/GNOME/mutter/-/issues/2559
Bug-Debian: https://bugs.debian.org/1024438
Signed-off-by: Simon McVittie <smcv@debian.org>
Forwarded: https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/2774

Gbp-Pq: Name workspace-Don-t-crash-on-invalid-argument-to-meta_workspa.patch